feat: #104898 检查试卷状态

+++ src/ApiBundle/Api/Resource/Assessment/Assessment.php
@@ -94,12 +94,15 @@ class Assessment extends AbstractResource
             throw CommonException::ERROR_PARAMETER_MISSING();
         }
         $assessment = $this->getAssessmentService()->getAssessment($assessmentId);
+        if (!in_array($assessment['status'], ['draft', 'closed', 'failure'])) {
+            throw AssessmentException::STATUS_ERROR();
+        }
         try {
             $this->biz['db']->beginTransaction();
-            $this->getAssessmentService()->deleteAssessment($assessmentId);
             if ('random' == $assessment['type']) {
                 $this->getAssessmentService()->deleteAssessmentByParentId($assessmentId);
             }
+            $this->getAssessmentService()->deleteAssessment($assessmentId);
             $this->biz['db']->commit();
         } catch (\Exception $e) {
             $this->biz['db']->rollback();
